git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@53 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CTextbox.cpp
@@ -51,7 +51,7 @@ void CTextbox::Draw(SDL_Surface *bmpDest)
 	if(iFlags & TXF_PASSWORD) {
 
 		// Draw astericks for password
-		for(i=0;i<strlen(sText);i++)
+		for(i=0;i<(int)strlen(sText);i++)
 			text[i]='*';
 		text[i+1] = '\0';
 	}
